refactor: add component for loader (ECOMM-72)
