+++ src/components/ContentLoaderWrapper/index.tsx
@@ -1,0 +1,3 @@
+import ContentLoaderWrapper from './ContentLoaderWrapper';
+
+export default ContentLoaderWrapper;
