+++ src/pages/CatalogPage/CatalogPage.tsx
@@ -1,44 +1,35 @@
-import { Box, CircularProgress, Grid, Typography } from '@mui/material';
+import { Box, Grid, Typography } from '@mui/material';
 import { getProducts } from 'services/sdk/product';
-import { useContext, useEffect, useState } from 'react';
-import AppContext from 'context';
+import { useMemo, useState } from 'react';
 import CatalogProductItem from 'components/CatalogProductItem';
 import { Product } from 'types/types';
+import ContentLoaderWrapper from 'components/ContentLoaderWrapper';
 
 function CatalogPage() {
-  const [isLoading, setIsLoading] = useState(true);
   const [products, setProducts] = useState<Product[]>([]);
 
-  const appContext = useContext(AppContext);
-  const setMessage = appContext?.setMessage;
+  const loadingLogic = useMemo(() => {
+    return async () => {
+      const productsData = await getProducts();
+      setProducts(productsData);
+    };
+  }, []);
 
-  useEffect(() => {
-    getProducts()
-      .then((productsData) => setProducts(productsData))
-      .catch((error) => {
-        if (setMessage)
-          setMessage({ severity: 'error', text: error instanceof Error ? error.message : 'Unknown error' });
-      })
-      .finally(() => setIsLoading(false));
-  }, [setMessage]);
-
-  return isLoading ? (
-    <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', height: '100%' }}>
-      <CircularProgress />
-    </Box>
-  ) : (
-    <Box px={3}>
-      <Typography component="h2" variant="h2" mb={2}>
-        Catalog
-      </Typography>
-      <Grid container spacing={3}>
-        {products.map((product) => (
-          <Grid item xs={12} sm={6} md={4} lg={3} key={product.id}>
-            <CatalogProductItem product={product} />
-          </Grid>
-        ))}
-      </Grid>
-    </Box>
+  return (
+    <ContentLoaderWrapper loadingLogic={loadingLogic}>
+      <Box px={3}>
+        <Typography component="h2" variant="h2" mb={2}>
+          Catalog
+        </Typography>
+        <Grid container spacing={3}>
+          {products.map((product) => (
+            <Grid item xs={12} sm={6} md={4} lg={3} key={product.id}>
+              <CatalogProductItem product={product} />
+            </Grid>
+          ))}
+        </Grid>
+      </Box>
+    </ContentLoaderWrapper>
   );
 }
